LPS-54221 move MethodSignature to portal-service
