+++ portal-impl/src/com/liferay/portal/monitoring/statistics/service/ServerStatistics.java
@@ -14,8 +14,8 @@
 
 package com.liferay.portal.monitoring.statistics.service;
 
+import com.liferay.portal.kernel.monitoring.MethodSignature;
 import com.liferay.portal.kernel.monitoring.statistics.DataSampleProcessor;
-import com.liferay.portal.monitoring.jmx.MethodSignature;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
