+++ portal-impl/src/com/liferay/portal/monitoring/statistics/service/ServiceMonitorAdvice.java
@@ -14,10 +14,10 @@
 
 package com.liferay.portal.monitoring.statistics.service;
 
+import com.liferay.portal.kernel.monitoring.MethodSignature;
 import com.liferay.portal.kernel.monitoring.RequestStatus;
 import com.liferay.portal.kernel.monitoring.statistics.DataSampleThreadLocal;
 import com.liferay.portal.kernel.util.AutoResetThreadLocal;
-import com.liferay.portal.monitoring.jmx.MethodSignature;
 import com.liferay.portal.spring.aop.ChainableMethodAdvice;
 
 import java.lang.reflect.Method;
